[PLUGINS] updated plugin module structure: implement #247
